Ajustes comentarios

+++ Guia2/src/guia2/SerieParImpar.java
@@ -6,7 +6,7 @@ public class SerieParImpar {
 
     public static void main(String[] args) {
         Scanner leer= new Scanner (System.in);
-        double x,y,mayor,menor;
+        int x,y,mayor,menor;
         byte opc;
         
         System.out.println("Bienvenido a su programa serie par o impar\nEste programa le permite ver la serie par o impar de dos numeros que usted ingrese");
@@ -17,7 +17,7 @@ public class SerieParImpar {
         
              
         System.out.println("Por favor digite el segundo número");
-        y = leer.nextDouble();
+        y = leer.nextInt();
        
         if (y>x){
                 mayor=y;
@@ -39,7 +39,7 @@ public class SerieParImpar {
                             System.out.println("su serie es:");
                             menor=menor+1;
                             
-                            for (double i = menor ; i <= mayor; i=i+2) {
+                            for (int i = menor ; i <= mayor; i=i+2) {
                                 System.out.print(i+" ");
                             }
                             System.out.println("");
@@ -49,7 +49,7 @@ public class SerieParImpar {
                             System.out.println("su serie es:");
                             menor=menor+0;
                             
-                            for (double i = menor ; i <= mayor; i=i+2) {
+                            for (int i = menor ; i <= mayor; i=i+2) {
                                 System.out.print(i+" ");
                             }
                             System.out.println("");
@@ -63,7 +63,7 @@ public class SerieParImpar {
                             System.out.println("y el número es par, por lo tanto.");
                             System.out.println("su serie es:");
                             menor=menor+1;
-                            for (double i = menor ; i <= mayor; i=i+2) {
+                            for (int i = menor ; i <= mayor; i=i+2) {
                             System.out.print(i+" ");
                             }
                             System.out.println("");
@@ -73,7 +73,7 @@ public class SerieParImpar {
                         System.out.println("su serie es:");
                             menor=menor+0;
                             
-                            for (double i = menor ; i <= mayor; i=i+2) {
+                            for (int i = menor ; i <= mayor; i=i+2) {
                                 System.out.print(i+" ");
                             }
                             System.out.println("");
